kimi/email_scrap/scripts/requirements.txt
# CRWM Email Scraping - Combined Dependencies
# ============================================
# This file combines all requirements from:
# - Main requirements.txt (automation + backend API)
# - FINAL-complete/requirements.txt (error dashboard)

# Core Python
python>=3.9

# ============================================
# Automation Pipeline Dependencies
# ============================================
# Browser Automation (Core)
playwright>=1.35.0

# Scraping & API
apify-client
curl-cffi>=0.5.0  # Faster requests with browser masquerading
beautifulsoup4>=4.12.0  # HTML parsing

# ============================================
# Backend API Dependencies (FastAPI)
# ============================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database (PostgreSQL - for backend API)
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Database (MySQL - for error dashboard)
pymysql==1.1.0

# Authentication & Security
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
bcrypt==5.0.0  # Using higher version (5.0.0 > 3.2.0)
passlib[bcrypt]==1.7.4  # Password hashing for error dashboard

# ============================================
# Error Dashboard Dependencies
# ============================================
aiofiles==23.2.1  # Async file operations
python-dotenv==1.0.0  # Environment variable management
jinja2==3.1.2  # Template engine for HTML rendering

# ============================================
# Installation Instructions
# ============================================
# 1. pip install -r requirements.txt
# 2. playwright install  (to install browser binaries)
# 3. Create .env file with database credentials (see README.md)
